news_bot/requirements.txt
requests
yfinance
feedparser
python-dotenv
google-genai
